+++ app/controllers/app.js
@@ -45,6 +45,25 @@ var tableCache = LRU({
   maxAge: global.settings.tableCacheMaxAge || 1000*60*10
 });
 
+function pad(n) { return n < 10 ? '0' + n : n }
+Date.prototype.toJSON = function() {
+  var s = this.getFullYear() + '-'
+      + pad(this.getMonth() + 1) + '-'
+      + pad(this.getDate()) + 'T'
+      + pad(this.getHours()) + ':'
+      + pad(this.getMinutes()) + ':'
+      + pad(this.getSeconds());
+  var offset = this.getTimezoneOffset();
+  if (offset == 0) s += 'Z';
+  else {
+    s += ( offset < 0 ? '+' : '-' )
+      + pad(Math.abs(offset / 60))
+      + pad(Math.abs(offset % 60))
+        
+  }
+  return s;
+}
+
 app.use(express.bodyParser());
 app.enable('jsonp callback');
 app.set("trust proxy", true);